+++ doc/changelog.md
@@ -4,8 +4,8 @@
 * **0.45.1 (2024-09-29)**:
   - Make copy docker-buildx binary to temporary config directory work on windows too ([1819](https://github.com/fabric8io/docker-maven-plugin/pull/1819))
   - Pull FROM images in relative path Dockerfiles ([1823](https://github.com/fabric8io/docker-maven-plugin/issues/1823))
-  - Docker-compose healthcheck configuration support
-  - Docker container wait timeout default value made configurable using defaultContainerWaitTimeout configuration option
+  - Docker-compose healthcheck configuration support ([1825](https://github.com/fabric8io/docker-maven-plugin/pull/1825))
+  - Docker container wait timeout default value made configurable using defaultContainerWaitTimeout configuration option ([1825](https://github.com/fabric8io/docker-maven-plugin/pull/1825))
 
 * **0.45.0 (2024-07-27)**:
   - Automatically create parent directories of portPropertyFile path ([1761](https://github.com/fabric8io/docker-maven-plugin/pull/1761))