+++ doc/changelog.md
@@ -1,5 +1,7 @@
 # ChangeLog
 
+* **0.47-SNAPSHOT**:
+
 * **0.46.0 (2025-04-06)**:
   - Docker-compose healthcheck configuration support ([1825](https://github.com/fabric8io/docker-maven-plugin/pull/1825))
   - Docker container wait timeout default value made configurable using startContainerWaitTimeout configuration option ([1825](https://github.com/fabric8io/docker-maven-plugin/pull/1825))